+++ ui-sketcher-webview/src/global.d.ts
@@ -1,5 +1,6 @@
 interface Window {
   displayPreviewOnStart?: boolean;
+  partialRenderEnabled?: boolean;
   previewUrl?: string;
   relativePath: string;
   workspaceId: string;